Vector-Embeddings/Images/requirements.txt
numpy
pandas
streamlit
torch
Pillow
requests
scikit-learn
umap-learn
matplotlib
transformers==4.18.0
plotly
datasets
tqdm
